pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[bey]~=1.3",
    "livekit-plugins-openai>=0.10", # Version critical for Mini support
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "requests",
]

